Move backend folder to src/api
